Bug fixes to get naive MCTS working
This had been broken with the implementations of the Rule set up, and
was needed to re-run early experiments to confirm results.
Key changes are:
- ExpansionPolicy updated to have a separate createRoot() method
distinct from createNode(). The difference being that createRoot() does
not (cannot) have a parent MCTSNode.
- MCTSNode unexpandedActions now uses the correction to .isLegal()
already used by the Rule/Policy stuff. Otherwise Discard options
are always ignored